Add location toggle for comment compose
Add a toggle to enable users to set the location of the comment compose
window relative to the comments thread.

I personally prefer having the compose window on the bottom of the
thread as this forces in a way the users to at least glance at the other
comments before submitting their own.